test: post bid test
